% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmn_mcmc_from_stan.R
\name{dmn_mcmc_from_stan}
\alias{dmn_mcmc_from_stan}
\title{Bayesian Learning of Dynamic Multilayer Networks}
\usage{
dmn_mcmc_from_stan(stan_fit, directed = FALSE, weighted = FALSE)
}
\arguments{
\item{stan_fit}{Object of class "stanfit".}

\item{directed}{Boolean. Indicates if the network is directed.}

\item{weighted}{Boolean. Indicates if the network is weighted.}
}
\value{
An object of class \code{dmn_mcmc}.
}
\description{
\code{DynMultiNet_bin} Implements model from Durante and Dunson, 2018
}
\details{
Transform the output from the stan implemetation of the Latent Network model into a dmn_mcmc class object.
}
\keyword{internal}
