% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMultiNet-package.R
\docType{package}
\name{DynMultiNet-package}
\alias{DynMultiNet-package}
\alias{DynMultiNet}
\title{The 'DynMultiNet' package.}
\description{
Bayesian Learning of Dynamic Multilayer Networks.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
