% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMultiNet_bin.R
\name{DynMultiNet_bin}
\alias{DynMultiNet_bin}
\title{Bayesian Learning of Dynamic Multilayer Networks with binary data}
\usage{
DynMultiNet_bin(net_data, pred_data = NULL, H_dim = 10, k_x = 0.1,
  k_mu = 0.1, k_p = 0.1, a_1 = 2, a_2 = 2.5, n_iter_mcmc = 1e+05,
  out_file = NULL, log_file = NULL, quiet_mcmc = FALSE, use_cpp = TRUE)
}
\arguments{
\item{net_data}{Data frame with network information}

\item{pred_data}{Data frame with linked predictors information}

\item{H_dim}{Latent space dimension}

\item{n_iter_mcmc}{number of iterations for the MCMC}

\item{out_file}{Indicates a file (.RData) where the output should be saved}

\item{log_file}{Indicates a file (.txt) where the log of the process should be saved}

\item{quiet_mcmc}{Silent mode, no progress update}

\item{use_cpp}{Boolean, indicates wheter to use Cpp routines for efficient computation}
}
\value{
A list with the following components:
\describe{
    \item{\code{theta_mcmc}}{Matrix with the chain of the parameters in the model.}
}
}
\description{
\code{DynMultiNet_bin} Implements model from Durante and Dunson, 2018
}
\details{
The model assumes a latent variable approach
   
   \code{net_data} must be a data frame with the following columns
       \describe{
       \item{\code{source}}{Start Node}
       \item{\code{target}}{End node}
       \item{\code{weight}}{Edge weight}
       \item{\code{time}}{time associated with the edge}
       \item{\code{layer}}{Layer associated with the edge}
       }
}
\examples{

   DynMultiNet_bin( net_data,
                    pred_data=NULL,
                    H_dim=10,
                    k_x=0.10, k_mu=0.10, k_p=0.10,
                    a_1=2, a_2=2.5,
                    n_iter_mcmc=100000,
                    out_file=NULL, log_file=NULL,
                    quiet_mcmc=FALSE )

}
