% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynMultiNet_bin_dir.R
\name{DynMultiNet_bin_dir}
\alias{DynMultiNet_bin_dir}
\title{Bayesian Learning of Dynamic Multilayer Networks with binary directed edges}
\usage{
DynMultiNet_bin_dir(net_data, pred_data = NULL, directed = TRUE,
  H_dim = 10, R_dim = 10, k_x = 0.1, k_mu = 0.1, k_p = 0.1, a_1 = 2,
  a_2 = 2.5, n_iter_mcmc = 10000, n_burn = n_iter_mcmc/2, n_thin = 3,
  out_file = NULL, log_file = NULL, quiet_mcmc = FALSE)
}
\arguments{
\item{net_data}{Data frame.Network information.}

\item{pred_data}{Data frame. Linked predictors information.}

\item{directed}{Boolean. Indicates if the provided network is directed, i.e. the adjacency matrix is assymetrical.}

\item{H_dim}{Integer. Latent space dimension.}

\item{R_dim}{Integer. Latent space dimension, for layer specific latent vectors.}

\item{n_iter_mcmc}{Integer. Number of iterations for the MCMC.}

\item{n_burn}{Integer. Number of iterations discarded as part of the MCMC warming up period at the beginning of the chain.}

\item{n_thin}{Integer. Number of iterations discarded for thining the chain (reducing the autocorrelation). We keep 1 of every n_thin iterations.}

\item{out_file}{String. Indicates a file (.RData) where the output should be saved.}

\item{log_file}{String. Indicates a file (.txt) where the log of the process should be saved.}

\item{quiet_mcmc}{Boolean. Indicates if silent mode is preferes, if \code{FALSE} progress update is displayed.}
}
\value{
A list with the following components:
\describe{
    \item{\code{theta_mcmc}}{Matrix with the chain of the parameters in the model.}
}
}
\description{
\code{DynMultiNet_bin} Implements model from Durante and Dunson, 2018
}
\details{
The model assumes a latent variable approach
   
   \code{net_data} must be a data frame with the following columns
       \describe{
       \item{\code{source}}{Start Node}
       \item{\code{target}}{End node}
       \item{\code{weight}}{Edge weight}
       \item{\code{time}}{time associated with the edge}
       \item{\code{layer}}{Layer associated with the edge}
       }
}
\examples{

   DynMultiNet_bin( net_data,
                    pred_data=NULL,
                    directed=TRUE,
                    H_dim=10, R_dim=10,
                    k_x=0.10, k_mu=0.10, k_p=0.10,
                    a_1=2, a_2=2.5,
                    n_iter_mcmc=10000, n_burn=n_iter_mcmc/2, n_thin=3,
                    out_file=NULL, log_file=NULL,
                    quiet_mcmc=FALSE )

}
