% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynNet_bin.R
\name{DynNet_bin}
\alias{DynNet_bin}
\title{Bayesian Learning of Dynamic Networks with binary data}
\usage{
DynNet_bin(net_data, H_dim = 10, k_x = 0.1, k_mu = 0.1, a_1 = 2,
  a_2 = 2.5, n_iter_mcmc = 100000, out_file = NULL, log_file = NULL,
  out_dir = NULL, quiet_mcmc = FALSE)
}
\arguments{
\item{net_data}{Data frame with network information}

\item{H_dim}{Latent space dimension}

\item{n_iter_mcmc}{number of iterations for the MCMC}

\item{out_file}{Indicates a file (.RData) where the output should be saved}

\item{log_file}{Indicates a file (.txt) where the log of the process should be saved}

\item{quiet_mcmc}{Silent mode, no progress update}
}
\value{
A list with the following components:
\describe{
    \item{\code{theta_mcmc}}{Matrix with the chain of the parameters in the model.}
}
}
\description{
\code{DynNet_bin} Implements model from Durante and Dunson, 2014
}
\details{
The model assumes a latent variable approach
   
   \code{net_data} must be a data frame with the following columns
       \describe{
       \item{\code{source}}{Start Node}
       \item{\code{target}}{End node}
       \item{\code{weight}}{Edge weight}
       \item{\code{time}}{time associated with the edge}
       \item{\code{layer}}{Layer associated with the edge}
       }
}
\examples{

   DynMultiNet_bin( net_data,
                    H_dim=10, k_x=0.10, k_mu=0.10, a_1=2, a_2=2.5,
                    n_iter_mcmc=100000,
                    out_file=NULL, log_file=NULL, out_dir=NULL,
                    quiet_mcmc=FALSE )

}
