% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linpred_s_ijtk.R
\name{get_linpred_s_ijtk}
\alias{get_linpred_s_ijtk}
\title{Bayesian Learning of Dynamic Multilayer Networks with binary data}
\usage{
get_linpred_s_ijtk(y_ijtk, mu_tk, x_iht, pred_all = NULL, layer_all = NULL,
  z_tp = NULL, z_tkp = NULL, z_ijtkp = NULL, beta_z_tp = NULL,
  beta_z_tkp = NULL, beta_z_ijtkp = NULL, pred_id_tp = NULL,
  pred_id_tkp = NULL, pred_id_ijtkp = NULL)
}
\arguments{
\item{y_ijtk}{Array of dimension dim(y_ijtk)=c(V_net,V_net,T_net,K_net). Binary undirected links between nodes i and j, at time t, on layer k}

\item{mu_tk}{Matrix. Edge specific baseline factor of the linear predictor}

\item{x_iht}{Array. Latent coordinates of element i in dimension h}

\item{z_tp}{Matrix. Global predictors data, predictor p at time t}

\item{z_tkp}{Array. Layer specific predictors data, predictor p at time t for layer k}

\item{z_ijtkp}{Array. Edge specific predictors data, predictor p at time t for edge bewtween i and j in layer k}

\item{beta_z_tp}{Column Matrix. Coefficients associated with z_tp}

\item{beta_z_tkp}{Column Matrix. Coefficients associated with z_tkp}

\item{beta_z_ijtkp}{Column Matrix. Coefficients associated with z_ijtkp}

\item{pred_id_tp}{Matrix. List of global predictors}

\item{pred_id_tkp}{Matrix. List of layer specific predictors}

\item{pred_id_ijtkp}{Matrix. List of edge specific predictors}
}
\value{
An array s_ijtk of dimension dim(s_ijtk)=c(V_net,V_net,T_net,K_net)
}
\description{
\code{DynMultiNet_bin} Implements model from Durante and Dunson, 2018
}
\details{
Linear predictor of response variable y_ijtk under a logistic link,
   logit( E(y_ijtk) ) = s_ijtk
}
\keyword{internal}
