% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linpred_s_ijtk.R
\name{get_linpred_s_ijtk}
\alias{get_linpred_s_ijtk}
\title{Bayesian Learning of Dynamic Multilayer Networks with binary data}
\usage{
get_linpred_s_ijtk(y_ijtk, mu_tk, x_iht_shared, x_ihtk = NULL,
  pred_all = NULL, layer_all = NULL, z_tkp = NULL, z_ijtkp = NULL,
  beta_z_layer = NULL, beta_z_edge = NULL, pred_id_layer = NULL,
  pred_id_edge = NULL)
}
\arguments{
\item{y_ijtk}{Array of dimension dim(y_ijtk)=c(V_net,V_net,T_net,K_net). Binary undirected links between nodes i and j, at time t, on layer k}

\item{mu_tk}{Matrix. Edge specific baseline factor of the linear predictor}

\item{x_iht_shared}{Array. Global latent coordinates of element i in dimension h}

\item{x_ihtk}{Array. Layer Specific Latent coordinates of element i in dimension h. Only for multilayer networks.}

\item{pred_all}{Vector. Names of all predictors.}

\item{layer_all}{Vector. Names of all layers.}

\item{z_tkp}{Array. Layer specific predictors data, predictor p at time t for layer k}

\item{z_ijtkp}{Array. Edge specific predictors data, predictor p at time t for edge bewtween i and j in layer k}

\item{beta_z_layer}{Column Matrix. Coefficients associated with z_tkp}

\item{beta_z_edge}{Column Matrix. Coefficients associated with z_ijtkp}

\item{pred_id_layer}{Matrix. List of layer specific predictors}

\item{pred_id_edge}{Matrix. List of edge specific predictors}
}
\value{
An array s_ijtk of dimension dim(s_ijtk)=c(V_net,V_net,T_net,K_net)
}
\description{
\code{DynMultiNet_bin} Implements model from Durante and Dunson, 2018
}
\details{
Linear predictor of response variable y_ijtk under a logistic link,
   logit( E(y_ijtk) ) = s_ijtk
}
\keyword{internal}
