% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dmn_mcmc.R
\name{plot_dmn_mcmc}
\alias{plot_dmn_mcmc}
\title{Plotting posterior MCMC results for the function \code{dmn_sampling}}
\usage{
plot_dmn_mcmc(x, param = c("pi_ijtk", "mu_tk", "x_ith_shared", "x_ithk",
  "tau_h_shared", "tau_h_k", "beta")[1], node_i = NULL, node_j = NULL,
  layer_k = NULL, h = NULL, pred_p = NULL, lat_space = NULL,
  cred_int_type = c("empirical", "norm")[1], n_burn = 0, n_thin = 1)
}
\arguments{
\item{x}{an object of class "\code{plot_dmn_mcmc}"}

\item{param}{Character. Specifies the parameter that will be plotted. Possible parameters are:
\describe{
    \item{\code{"pi_ijtk"}}{Edge probabilities between node_i and node_j at time t in layer_k.}
    \item{\code{"mu_tk"}}{Baseline process for edge probability at layer_k.}
    \item{\code{"x_ith_shared"}}{Shared latent coordinates for node_i in dimension h, corresponding to edge probability.}
    \item{\code{"x_ithk"}}{Layer-specific latent coordinates for node_i at layer_k in dimension h, corresponding to edge probability.}
    \item{\code{"tau_h_shared"}}{MCMC chain for the shrinkage parameter of the global latent space in dimension h, corresponding to edge probability.}
    \item{\code{"tau_h_k"}}{MCMC chain for the shrinkage parameter of the layer-specific latent space in dimension h, corresponding to edge probability.}
    \item{\code{"r_ijtk"}}{Expected value of edge weight between node_i and node_j at time t in layer_k.}
    \item{\code{"sigma_w_k"}}{Variance of edge weight at layer_k.}
    \item{\code{"lambda_tk_mcmc"}}{Baseline process for edge weight at layer_k.}
    \item{\code{"u_ith_shared"}}{Shared latent coordinates for node_i in dimension h, corresponding to edge weight.}
    \item{\code{"x_ithk"}}{Layer-specific latent coordinates for node_i at layer_k in dimension h, corresponding to edge weight}
    \item{\code{"rho_h_shared"}}{MCMC chain for the shrinkage parameter of the global latent space in dimension h, corresponding to edge weight}
    \item{\code{"rho_h_k"}}{MCMC chain for the shrinkage parameter of the layer-specific latent space in dimension h, corresponding to edge weight}
    }}

\item{node_i}{character/numeric. Id of a node in the network.}

\item{node_j}{character/numeric. Id of a node in the network.}

\item{layer_k}{character/numeric. Id of a layer in the network.}

\item{h}{numeric. Dimension of the latent space.}

\item{pred_p}{character/numeric. Id of a predictor in the exogenous data.}

\item{lat_space}{character In a directed network, specifies if the latent coordinate is in the "send" or "receive" space.}

\item{cred_int_type}{character. Type of credible intervals: Empirical or normal approximation}

\item{n_burn}{Integer. Number of iterations discarded as part of the MCMC warming up period at the beginning of the chain.}

\item{n_thin}{Integer. Number of iterations discarded for thining the chain (reducing the autocorrelation). We keep 1 of every n_thin iterations.}
}
\description{
Plotting method for objects inheriting from class "\code{dmn_mcmc}".
}
\examples{

\dontrun{

plot_dmn_mcmc( x=dmn_mcmc,
               param="pi_ijtk",
               node_i=dmn_mcmc$node_all[2],
               node_j=dmn_mcmc$node_all[1],
               layer_k=1 )
               
plot_dmn_mcmc( x=dmn_mcmc,
               param="mu_tk",
               layer_k=dmn_mcmc$layer_all[1] )
               
plot_dmn_mcmc( x=dmn_mcmc,
               param="x_ith_shared",
               node_i=dmn_mcmc$node_all[1],
               h=1 )

plot_dmn_mcmc( x=dmn_mcmc,
               param="x_ithk",
               node_i=dmn_mcmc$node_all[1],
               h=1, layer_k=1 )
}
}
